chore: split stale symbol surfaces script and mark executable
Extract models, visitors, and analysis into helper modules so each file
stays under the 600-line gate. Set executable bit on the entry script
for the shebang pre-commit check.

Made-with: Cursor
